move http functionality into a separate folder
