+++ core/snow.c
@@ -25,17 +25,17 @@
 #include "core/conf.h"
 #include "core/connection.h"
 #include "core/core.h"
-#include "core/http.h"
 #include "core/log.h"
 #include "core/module.h"
 #include "core/msg.h"
-#include "json-c/json.h"
 #include "core/peer.h"
 #include "core/roominfo.h"
 #include "core/snow.h"
 #include "core/snw_event.h"
 #include "core/task.h"
 #include "core/utils.h"
+#include "http/http.h"
+#include "json-c/json.h"
 #include "websocket/websocket.h"
 
 int
